automation/requirements.txt
pymongo==4.4.1
yagmail==0.14.245
python-dotenv==1.0.0